requirements.txt
scikit-image
scikit-fmm
matplotlib
numpy
imageio
PyQt5
pyyaml